Refactor domain info models
