+++ epp-client/src/epp.rs
@@ -10,7 +10,6 @@ pub use request::contact::create::*;
 pub use request::contact::delete::*;
 pub use request::contact::info::*;
 pub use request::contact::update::*;
-pub use request::domain::info::*;
 pub use request::domain::renew::*;
 pub use request::domain::rgp::report::*;
 pub use request::domain::rgp::request::*;
@@ -29,7 +28,6 @@ pub use response::contact::create::*;
 pub use response::contact::delete::*;
 pub use response::contact::info::*;
 pub use response::contact::update::*;
-pub use response::domain::info::*;
 pub use response::domain::renew::*;
 pub use response::domain::rgp::report::*;
 pub use response::domain::rgp::request::*;
