+++ epp-client/src/epp/request/domain.rs
@@ -1,6 +1,5 @@
 //! Types for EPP domain requests
 
-pub mod info;
 pub mod renew;
 pub mod rgp;
 pub mod transfer;
