+++ epp-client/src/epp/response/domain.rs
@@ -1,6 +1,5 @@
 //! Types for EPP domain responses
 
-pub mod info;
 pub mod renew;
 pub mod rgp;
 pub mod transfer;
